requirements.txt
numpy==1.19.2
Pillow==8.3.1
